server/requirements.txt
#cognitive-services-speech-sdk
langchain
langchain-openai
langchain-core
langchain_community
docx2txt
langchain_chroma
python-multipart
python-dotenv
pydub
uvicorn[standard] 
redis
jinja2
azure-identity
azure-cosmos
azure.cognitiveservices.speech
azure-search-documents
azure-monitor-opentelemetry
opentelemetry-instrumentation-langchain